backend/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
python-multipart
ollama